refactor: split common.py into package, fix tests, improve gov_rules filtering
- Split scripts/common.py into scripts/common/ package (8 submodules + __init__.py)
- Add backward-compatible aliases for _CacheManager, _RateLimitMode, etc.
- Add create_crawler_headers() factory function, unified across 6 crawlers
- Fix test_article_search.py mock paths (download.* -> article_search.*)
- Fix gov_rules_crawler.py keyword filtering (server-side Athena API)
- Add --rps flag (replaces --rate-limit N numeric parsing)
- Add requirements-dev.txt (pytest, coverage)
- Add *_output/logs/ to .gitignore
- Fix type annotation warnings in download.py
