+++ requirements-dev.txt
@@ -1,0 +1,3 @@
+# Development dependencies
+pytest>=7.0.0
+coverage>=7.0.0
