+++ scripts/download.py
@@ -18,7 +18,7 @@ Rate Limiting:
   --rate-limit off       No throttling (small tasks)
   --rate-limit fixed     Fixed ~5 req/s (medium tasks)
   --rate-limit adaptive  Adaptive speed (large tasks)
-  --rate-limit N         Fixed N requests per second (e.g., --rate-limit 3)
+  --rps N                Fixed N requests per second (e.g., --rps 3)
 
 Environment:
   SSL verification is disabled by default. To enable: export NPC_LAW_VERIFY_SSL=1
@@ -39,25 +39,27 @@ import json
 import os
 import re
 import sys
-from pathlib import Path
 
 from common import (
+    DEFAULT_USER_AGENT,
     _CacheManager,
     _RateLimitConfig,
     _RateLimitMode,
     _SmartRateLimiter,
-    _backoff,
-    chinese_to_int as _chinese_to_int,
-    DEFAULT_USER_AGENT,
-    ensure_dir,
-    extract_article_number as _extract_article_number,
-    extract_paragraphs_from_docx as _extract_paragraphs_from_docx,
     get_cache,
     http_request,
     init_limiter,
-    int_to_chinese as _int_to_chinese,
+)
+from common import (
+    extract_paragraphs_from_docx as _extract_paragraphs_from_docx,
+)
+from common import (
     is_article_line as _is_article_line,
+)
+from common import (
     match_article_query as _match_article_query,
+)
+from common import (
     split_into_articles as _split_into_articles,
 )
 
@@ -90,33 +92,52 @@ def _request(method, url, **kwargs):
 # ---------------------------------------------------------------------------
 
 
-def search_laws(keyword: str, page: int = 1, size: int = 20,
-                search_range: int = 1, search_type: int = 2,
-                status_filter: list = None) -> dict:
+def search_laws(
+    keyword: str,
+    page: int = 1,
+    size: int = 20,
+    search_range: int = 1,
+    search_type: int = 2,
+    status_filter: list[int] | None = None,
+) -> dict:
     """Search laws. search_range: 1=title, 2=content."""
-    cache_key = _cache._key("search", keyword, str(page), str(size),
-                            str(search_range), str(search_type),
-                            "".join(str(s) for s in (status_filter or [])))
+    cache_key = _cache._key(
+        "search",
+        keyword,
+        str(page),
+        str(size),
+        str(search_range),
+        str(search_type),
+        "".join(str(s) for s in (status_filter or [])),
+    )
     cached = _cache.get(cache_key, max_age=3600)
     if cached:
         return cached
 
     payload = {
         "searchRange": search_range,
-        "sxrq": [], "gbrq": [],
+        "sxrq": [],
+        "gbrq": [],
         "sxx": status_filter or [],
         "searchType": search_type,
         "xgzlSearch": False,
         "searchContent": keyword,
         "orderByParam": {"order": "-1", "sort": ""},
-        "flfgCodeId": [], "zdjgCodeId": [], "gbrqYear": [],
+        "flfgCodeId": [],
+        "zdjgCodeId": [],
+        "gbrqYear": [],
         "pageNum": page,
         "pageSize": size,
     }
     resp = _request(
-        "POST", f"{BASE_URL}/law-search/search/list",
+        "POST",
+        f"{BASE_URL}/law-search/search/list",
         json=payload,
-        headers={**HEADERS, "Referer": f"{BASE_URL}/search", "Content-Type": "application/json"},
+        headers={
+            **HEADERS,
+            "Referer": f"{BASE_URL}/search",
+            "Content-Type": "application/json",
+        },
     )
     data = resp.json()
     _cache.set(cache_key, data)
@@ -129,7 +150,9 @@ def fetch_detail(bbbs_id: str) -> dict:
     if cached:
         return cached
     try:
-        resp = _request("GET", f"{BASE_URL}/law-search/search/flfgDetails", params={"bbbs": bbbs_id})
+        resp = _request(
+            "GET", f"{BASE_URL}/law-search/search/flfgDetails", params={"bbbs": bbbs_id}
+        )
         data = resp.json() if resp.status_code == 200 else {}
         _cache.set(cache_key, data)
         return data
@@ -153,16 +176,23 @@ def parse_detail(data: dict) -> dict:
         "effective_date": d.get("sxrq", ""),
         "status_code": sxx_code,
         "status_str": sxx_to_str(sxx_code),
-        "word_url": f"{BASE_URL}/{oss['ossWordPath']}" if oss.get("ossWordPath") else None,
+        "word_url": f"{BASE_URL}/{oss['ossWordPath']}"
+        if oss.get("ossWordPath")
+        else None,
         "pdf_url": f"{BASE_URL}/{oss['ossPdfPath']}" if oss.get("ossPdfPath") else None,
-        "word_ofd_url": f"{BASE_URL}/{oss['ossWordOfdPath']}" if oss.get("ossWordOfdPath") else None,
-        "pdf_ofd_url": f"{BASE_URL}/{oss['ossPdfOfdPath']}" if oss.get("ossPdfOfdPath") else None,
+        "word_ofd_url": f"{BASE_URL}/{oss['ossWordOfdPath']}"
+        if oss.get("ossWordOfdPath")
+        else None,
+        "pdf_ofd_url": f"{BASE_URL}/{oss['ossPdfOfdPath']}"
+        if oss.get("ossPdfOfdPath")
+        else None,
     }
 
 
 def get_download_url(bbbs_id: str, fmt: str = "docx") -> str:
     resp = _request(
-        "GET", f"{BASE_URL}/law-search/download/pc",
+        "GET",
+        f"{BASE_URL}/law-search/download/pc",
         params={"format": fmt, "bbbs": bbbs_id},
         headers={**HEADERS, "Referer": f"{BASE_URL}/detail?id={bbbs_id}"},
     )
@@ -175,7 +205,7 @@ def get_download_url(bbbs_id: str, fmt: str = "docx") -> str:
     return url
 
 
-def download_file(url: str, output_path: str = None) -> str:
+def download_file(url: str, output_path: str | None = None) -> str:
     resp = _request("GET", url)
     resp.raise_for_status()
     ct = resp.headers.get("Content-Type", "")
@@ -185,7 +215,10 @@ def download_file(url: str, output_path: str = None) -> str:
         cd = resp.headers.get("Content-Disposition", "")
         m = re.search(r'filename="?([^"]+)"?', cd)
         output_path = m.group(1) if m else url.split("/")[-1].split("?")[0]
-    os.makedirs(os.path.dirname(os.path.abspath(output_path)) if os.path.dirname(output_path) else ".", exist_ok=True)
+    assert output_path is not None  # guaranteed by logic above
+    parent = os.path.dirname(os.path.abspath(output_path))
+    if parent:
+        os.makedirs(parent, exist_ok=True)
     with open(output_path, "wb") as f:
         f.write(resp.content)
     print(f"Downloaded: {output_path} ({len(resp.content)} bytes)")
@@ -257,7 +290,12 @@ def collect_search_urls(data: dict, fmt: str = "docx") -> list:
 
 
 def _detect_numbering_patterns(content_tree: dict) -> dict:
-    patterns = {"primary": "chinese", "has_chinese": False, "has_arabic": False, "sample_titles": []}
+    patterns = {
+        "primary": "chinese",
+        "has_chinese": False,
+        "has_arabic": False,
+        "sample_titles": [],
+    }
     if not content_tree:
         return patterns
 
@@ -319,9 +357,15 @@ def preview_law(bbbs_id: str):
     print(f"Category: {info['category']} | Authority: {info['authority']}")
     print(f"Publish: {info['publish_date']} | Status: {info['status_str']}")
     print(f"Total paragraphs: {len(paragraphs)} | Articles: {article_count}")
-    label_map = {"chinese": "中文数字 (如: 第一条)", "arabic": "阿拉伯数字 (如: 第1条)", "mixed": "混合"}
-    print(f"Numbering: {label_map.get(detected['primary'], detected['primary'])} | "
-          f"Samples: {', '.join(detected['sample_titles'][:3]) or 'N/A'}")
+    label_map = {
+        "chinese": "中文数字 (如: 第一条)",
+        "arabic": "阿拉伯数字 (如: 第1条)",
+        "mixed": "混合",
+    }
+    print(
+        f"Numbering: {label_map.get(detected['primary'], detected['primary'])} | "
+        f"Samples: {', '.join(detected['sample_titles'][:3]) or 'N/A'}"
+    )
     print()
     for num, text in articles[:20]:
         preview = text.replace("\n", " ")[:70]
@@ -331,7 +375,7 @@ def preview_law(bbbs_id: str):
         print(f"... ({len(articles) - 20} more)")
 
 
-def query_article(bbbs_id: str, query: str = None, grep: str = None):
+def query_article(bbbs_id: str, query: str | None = None, grep: str | None = None):
     raw = fetch_detail(bbbs_id)
     content_tree = (raw or {}).get("data", {}).get("content", {})
     detected = _detect_numbering_patterns(content_tree)
@@ -352,7 +396,7 @@ def query_article(bbbs_id: str, query: str = None, grep: str = None):
                 results.append((num, text))
         if not results:
             for num, text in articles:
-                if query in text[:len(query) + 15]:
+                if query in text[: len(query) + 15]:
                     results.append((num, text))
     else:
         print("Error: Use --article with query or --grep KEYWORD", file=sys.stderr)
@@ -381,46 +425,73 @@ def build_parser() -> argparse.ArgumentParser:
     parser = argparse.ArgumentParser(
         prog="download.py",
         description="Download laws/regulations from China's National Laws and Regulations Database.",
-        epilog=
-            "Examples:\n"
-            "  python download.py --search '出租车'\n"
-            "  python download.py --search '违约金' --range content --size 50\n"
-            "  python download.py --search '出租车' --urls-only --size 100 --rate-limit adaptive\n"
-            "  python download.py --info 2c909fdd678bf17901678bf73ebd064f\n"
-            "  python download.py --download ff80808172b5f24f0172d9f04f0910af --format docx out.doc\n"
-            "  python download.py --preview 2c909fdd678bf17901678bf74d7106b3\n"
-            "  python download.py --article 2c909fdd678bf17901678bf74d7106b3 '第三十八条'\n"
-            "\nRate Limiting:\n"
-            "  --rate-limit auto      Auto by task size (default: small=off, medium=fixed, large=adaptive)\n"
-            "  --rate-limit off       No throttling\n"
-            "  --rate-limit fixed     Fixed ~5 req/s\n"
-            "  --rate-limit adaptive  Adaptive speed with 429 backoff\n"
-            "  --rate-limit N         Fixed N req/s (e.g., --rate-limit 2)\n"
-            "\nEnvironment:\n"
-            "  NPC_LAW_VERIFY_SSL=1  Enable SSL verification (disabled by default)\n"
-            "  NPC_LAW_NO_CACHE=1    Disable cache",
+        epilog="Examples:\n"
+        "  python download.py --search '出租车'\n"
+        "  python download.py --search '违约金' --range content --size 50\n"
+        "  python download.py --search '出租车' --urls-only --size 100 --rate-limit adaptive\n"
+        "  python download.py --info 2c909fdd678bf17901678bf73ebd064f\n"
+        "  python download.py --download ff80808172b5f24f0172d9f04f0910af --format docx out.doc\n"
+        "  python download.py --preview 2c909fdd678bf17901678bf74d7106b3\n"
+        "  python download.py --article 2c909fdd678bf17901678bf74d7106b3 '第三十八条'\n"
+        "\nRate Limiting:\n"
+        "  --rate-limit auto      Auto by task size (default: small=off, medium=fixed, large=adaptive)\n"
+        "  --rate-limit off       No throttling\n"
+        "  --rate-limit fixed     Fixed ~5 req/s\n"
+        "  --rate-limit adaptive  Adaptive speed with 429 backoff\n"
+        "  --rps N                Fixed N req/s (e.g., --rps 3)\n"
+        "\nEnvironment:\n"
+        "  NPC_LAW_VERIFY_SSL=1  Enable SSL verification (disabled by default)\n"
+        "  NPC_LAW_NO_CACHE=1    Disable cache",
         formatter_class=argparse.RawDescriptionHelpFormatter,
     )
     parser.add_argument("--search", metavar="KEYWORD", help="Search by keyword")
-    parser.add_argument("--range", choices=["title", "content"], default="title",
-                        help="Search range: title (default) or content (full text)")
+    parser.add_argument(
+        "--range",
+        choices=["title", "content"],
+        default="title",
+        help="Search range: title (default) or content (full text)",
+    )
     parser.add_argument("--info", metavar="BBBS", help="Fetch detail/metadata")
     parser.add_argument("--download", metavar="BBBS", help="Download a file")
     parser.add_argument("--preview", metavar="BBBS", help="Preview law structure")
     parser.add_argument("--article", metavar="BBBS", help="Query specific article")
     parser.add_argument("--grep", metavar="KEYWORD", help="Search keyword in articles")
-    parser.add_argument("--page", type=int, default=1, help="Search page number (default: 1)")
-    parser.add_argument("--size", type=int, default=20, help="Search page size (default: 20)")
+    parser.add_argument(
+        "--page", type=int, default=1, help="Search page number (default: 1)"
+    )
+    parser.add_argument(
+        "--size", type=int, default=20, help="Search page size (default: 20)"
+    )
     parser.add_argument("--exact", action="store_true", help="Use exact title match")
-    parser.add_argument("--urls-only", action="store_true", help="Output signed URLs instead of downloading")
-    parser.add_argument("--format", default="docx", choices=["docx", "pdf"], help="Download format (default: docx)")
-    parser.add_argument("--status", help="Filter by status: 1=已废止,2=已修改,3=现行有效,4=尚未生效")
-    parser.add_argument("--rate-limit", default="auto",
-                        help="Rate limit mode (auto/off/fixed/adaptive/N)")
+    parser.add_argument(
+        "--urls-only",
+        action="store_true",
+        help="Output signed URLs instead of downloading",
+    )
+    parser.add_argument(
+        "--format",
+        default="docx",
+        choices=["docx", "pdf"],
+        help="Download format (default: docx)",
+    )
+    parser.add_argument(
+        "--status", help="Filter by status: 1=已废止,2=已修改,3=现行有效,4=尚未生效"
+    )
+    parser.add_argument(
+        "--rate-limit", default="auto", help="Rate limit mode (auto/off/fixed/adaptive)"
+    )
+    parser.add_argument(
+        "--rps",
+        type=float,
+        default=None,
+        help="Fixed requests per second (e.g., --rps 3), overrides --rate-limit",
+    )
     parser.add_argument("--no-cache", action="store_true", help="Disable cache")
     parser.add_argument("--cache-stats", action="store_true", help="Show cache stats")
     parser.add_argument("--cache-clear", action="store_true", help="Clear cache")
-    parser.add_argument("output", nargs="?", help="Output path for --download or direct URL mode")
+    parser.add_argument(
+        "output", nargs="?", help="Output path for --download or direct URL mode"
+    )
     return parser
 
 
@@ -431,24 +502,25 @@ def main():
     # Initialize rate limiter
     limiter, forced_mode = init_limiter("auto")
 
-    rl = (args.rate_limit or "auto").lower().strip()
-    if rl in ("auto", ""):
-        forced_mode = None
-    elif rl in ("off", "0", "no", "false", "none"):
-        forced_mode = _RateLimitMode.OFF
-    elif rl in ("fixed", "fix"):
+    if args.rps is not None:
+        cfg = _RateLimitConfig(fixed_rps=args.rps)
+        limiter = _SmartRateLimiter(cfg)
         forced_mode = _RateLimitMode.FIXED
-    elif rl in ("adaptive", "adapt"):
-        forced_mode = _RateLimitMode.ADAPTIVE
     else:
-        try:
-            rps = float(rl)
-            cfg = _RateLimitConfig(fixed_rps=rps)
-            limiter = _SmartRateLimiter(cfg)
+        rl = (args.rate_limit or "auto").lower().strip()
+        if rl in ("auto", ""):
+            forced_mode = None
+        elif rl in ("off", "0", "no", "false", "none"):
+            forced_mode = _RateLimitMode.OFF
+        elif rl in ("fixed", "fix"):
             forced_mode = _RateLimitMode.FIXED
-        except ValueError:
-            print(f"Warning: Unknown --rate-limit '{args.rate_limit}', using auto",
-                  file=sys.stderr)
+        elif rl in ("adaptive", "adapt"):
+            forced_mode = _RateLimitMode.ADAPTIVE
+        else:
+            print(
+                f"Warning: Unknown --rate-limit '{args.rate_limit}', using auto",
+                file=sys.stderr,
+            )
 
     estimated = _SmartRateLimiter.estimate_task_size(
         search=args.search,
@@ -461,8 +533,10 @@ def main():
     )
     mode = limiter.init_for_task(estimated, forced_mode=forced_mode)
     if mode != _RateLimitMode.OFF:
-        print(f"[RateLimit] {estimated} requests estimated -> {limiter.mode_desc()}",
-              file=sys.stderr)
+        print(
+            f"[RateLimit] {estimated} requests estimated -> {limiter.mode_desc()}",
+            file=sys.stderr,
+        )
 
     # Cache management
     if args.no_cache:
@@ -492,7 +566,9 @@ def main():
                 parser.print_help()
                 sys.exit(1)
             if not args.output.startswith(("http://", "https://")):
-                parser.error("Direct mode requires a URL starting with http:// or https://")
+                parser.error(
+                    "Direct mode requires a URL starting with http:// or https://"
+                )
             download_file(args.output, None)
             return
 
@@ -500,9 +576,14 @@ def main():
             raw = fetch_detail(args.info)
             print_detail(parse_detail(raw))
         elif args.search:
-            data = search_laws(args.search, page=args.page, size=args.size,
-                               search_range=search_range, search_type=search_type,
-                               status_filter=status_filter)
+            data = search_laws(
+                args.search,
+                page=args.page,
+                size=args.size,
+                search_range=search_range,
+                search_type=search_type,
+                status_filter=status_filter,
+            )
             if args.urls_only:
                 enriched = collect_search_urls(data, args.format)
                 json.dump(enriched, sys.stdout, ensure_ascii=False, indent=2)
