+++ scripts/gov_policy_library.py
@@ -19,9 +19,9 @@ import re
 from pathlib import Path
 
 from common import (
-    DEFAULT_USER_AGENT,
     _CacheManager,
     clean_text,
+    create_crawler_headers,
     ensure_dir,
     get_cache,
     http_request,
@@ -42,10 +42,7 @@ except ImportError:
 
 BASE_URL = "https://sousuo.www.gov.cn"
 SEARCH_API = "/search-gov/data"
-HEADERS = {
-    "User-Agent": DEFAULT_USER_AGENT,
-    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
-}
+HEADERS = create_crawler_headers()
 
 # Category filter mapping (from the page UI)
 CATEGORY_MAP = {
