+++ scripts/moj_law_crawler.py
@@ -21,9 +21,9 @@ from pathlib import Path
 from urllib.parse import urljoin
 
 from common import (
-    DEFAULT_USER_AGENT,
     _CacheManager,
     clean_text,
+    create_crawler_headers,
     ensure_dir,
     get_cache,
     http_request,
@@ -44,10 +44,7 @@ except ImportError:
 
 BASE_URL = "https://xzfg.moj.gov.cn"
 SEARCH_URL = f"{BASE_URL}/search2.html"
-HEADERS = {
-    "User-Agent": DEFAULT_USER_AGENT,
-    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
-}
+HEADERS = create_crawler_headers()
 
 STATUS_MAP = {
     "all": "",
