+++ scripts/party_law_crawler.py
@@ -23,9 +23,9 @@ from pathlib import Path
 from urllib.parse import urljoin
 
 from common import (
-    DEFAULT_USER_AGENT,
     _CacheManager,
     clean_text,
+    create_crawler_headers,
     ensure_dir,
     get_cache,
     http_request,
@@ -61,10 +61,7 @@ CATEGORY_MAP = {
     "党的监督保障法规": "jdbz",
 }
 
-HEADERS = {
-    "User-Agent": DEFAULT_USER_AGENT,
-    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
-}
+HEADERS = create_crawler_headers()
 
 _cache = get_cache("party-law-db")
