+++ tests/test_article_search.py
@@ -17,13 +17,13 @@ class TestSearchArticles(unittest.TestCase):
     @mock.patch("article_search._request")
     @mock.patch("article_search.get_download_url")
     @mock.patch("article_search.search_laws")
-    def test_single_match(self, mock_search, mock_get_url, mock_request, mock_cache, mock_extract):
+    def test_single_match(
+        self, mock_search, mock_get_url, mock_request, mock_cache, mock_extract
+    ):
         mock_search.return_value = {
             "code": 200,
             "total": 1,
-            "rows": [
-                {"bbbs": "id1", "title": "Test Law", "sxx": 3}
-            ],
+            "rows": [{"bbbs": "id1", "title": "Test Law", "sxx": 3}],
         }
         mock_cache.get_file.return_value = b"fake docx bytes"
         mock_extract.return_value = [
@@ -43,9 +43,11 @@ class TestSearchArticles(unittest.TestCase):
     @mock.patch("article_search.extract_paragraphs_from_docx")
     @mock.patch("article_search._cache")
     @mock.patch("article_search._request")
-    @mock.patch("download.get_download_url")
-    @mock.patch("download.search_laws")
-    def test_context(self, mock_search, mock_get_url, mock_request, mock_cache, mock_extract):
+    @mock.patch("article_search.get_download_url")
+    @mock.patch("article_search.search_laws")
+    def test_context(
+        self, mock_search, mock_get_url, mock_request, mock_cache, mock_extract
+    ):
         mock_search.return_value = {
             "code": 200,
             "total": 1,
@@ -58,7 +60,9 @@ class TestSearchArticles(unittest.TestCase):
             "第三条 内容三",
         ]
 
-        results = art_search.search_articles("违约金", max_laws=1, context=1, json_output=False)
+        results = art_search.search_articles(
+            "违约金", max_laws=1, context=1, json_output=False
+        )
 
         self.assertEqual(len(results[0]["articles"]), 3)  # 第二条 + 上下文
         nums = [a["article_num"] for a in results[0]["articles"]]
@@ -67,9 +71,11 @@ class TestSearchArticles(unittest.TestCase):
     @mock.patch("article_search.extract_paragraphs_from_docx")
     @mock.patch("article_search._cache")
     @mock.patch("article_search._request")
-    @mock.patch("download.get_download_url")
-    @mock.patch("download.search_laws")
-    def test_max_laws_limit(self, mock_search, mock_get_url, mock_request, mock_cache, mock_extract):
+    @mock.patch("article_search.get_download_url")
+    @mock.patch("article_search.search_laws")
+    def test_max_laws_limit(
+        self, mock_search, mock_get_url, mock_request, mock_cache, mock_extract
+    ):
         mock_search.return_value = {
             "code": 200,
             "total": 3,
